feat(*): move api tests to endpoints route
